+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Model file is no longer required (module checks the capabilities reported by the unit)
+* Added support for setting airflow direction (setAirDirH and setAirDirV)
+
+
 ## 0.1.1
 
 Correct the way model file is referenced.